@@ -30,29 +30,33 @@
     "trailingComma": "none"
   },
   "dependencies": {
-    "bulma": "^0.9.2",
-    "gray-matter": "^4.0.2",
-    "next": "^10.1.3",
+    "bootstrap": "^5.0.2",
+    "gray-matter": "^4.0.3",
+    "next": "^11.0.1",
+    "next-transpile-modules": "^8.0.0",
     "raw-loader": "^4.0.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-markdown": "^5.0.3",
+    "react-github-btn": "^1.2.0",
+    "react-markdown": "^6.0.2",
     "react-syntax-highlighter": "^15.4.3",
+    "react-typed": "^1.2.0",
+    "rehype-raw": "^5.1.0",
     "remark-gfm": "^1.0.0",
-    "sass": "^1.32.8",
+    "sass": "^1.35.1",
     "semver": "^7.3.5"
   },
   "devDependencies": {
     "@types/glob": "^7.1.3",
-    "@types/node": "^14.14.39",
-    "@types/react": "^17.0.3",
-    "@types/react-syntax-highlighter": "^13.5.0",
-    "@types/semver": "^7.3.4",
+    "@types/node": "^15.12.5",
+    "@types/react": "^17.0.11",
+    "@types/react-syntax-highlighter": "^13.5.1",
+    "@types/semver": "^7.3.6",
     "@types/webpack": "^5.28.0",
     "@types/webpack-env": "^1.16.0",
     "cross-env": "^7.0.3",
-    "firebase-tools": "^9.10.0",
-    "prettier": "^2.2.1",
-    "typescript": "^4.2.4"
+    "firebase-tools": "^9.14.0",
+    "prettier": "^2.3.2",
+    "typescript": "^4.3.4"
   }
 }
