@@ -1,18 +1,24 @@
+import Link from 'next/link';
 import { FunctionComponent } from 'react';
 
-const ContactBanner: FunctionComponent<{
-  title?: string;
-  ctaText?: string;
-}> = function (props) {
+const ContactBanner: FunctionComponent<{}> = function (props) {
   return (
-    <div className='section is-light'>
-      <div className='has-text-centered'>
-        <div className='title is-size-4'>
-          {props.title || 'Feedback? Questions? We are always here to help.'}
+    <div className='bg-gray-200'>
+      <div className='container py-8 border-bottom border-gray-300'>
+        <div className='row align-items-center text-center text-lg-start'>
+          <div className='col-12 col-lg-9'>
+            <h4 className='mb-4 mb-lg-1 fw-bold text-gray-700'>
+              Feedback? Questions? We are always here to help.
+            </h4>
+          </div>
+          <div className='col-12 col-lg-3 d-flex justify-content-end'>
+            <Link href='/contact/'>
+              <button className='col-12 btn btn-primary-soft'>
+                <span> Contact us</span>
+              </button>
+            </Link>
+          </div>
         </div>
-        <a href='/contact/' className='button is-link is-light'>
-          {props.ctaText || 'Contact us'}
-        </a>
       </div>
     </div>
   );
