@@ -1,9 +1,11 @@
 import { Children, createElement, FunctionComponent } from 'react';
-import ReactMarkdownWithHtml from 'react-markdown/with-html';
+import ReactMarkdown from 'react-markdown';
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
+import { vscDarkPlus } from 'react-syntax-highlighter/dist/cjs/styles/prism';
+import rehypeRaw from 'rehype-raw';
 import gfm from 'remark-gfm';
 import { linkTarget, transformLinkUri } from '../utils/url';
-import Blockquote from './blockquote';
-import CodeHighlighter from './code-highlighter';
+import Quotation from './quotation';
 import Quote from './quote';
 
 const flatten = (text, child) => {
@@ -12,36 +14,63 @@ const flatten = (text, child) => {
     : Children.toArray(child.props.children).reduce(flatten, text);
 };
 
+const heading = (props) => {
+  var children = Children.toArray(props.children);
+  var text = children.reduce(flatten, '');
+  var slug = text.toLowerCase().trim().replace(/\W/g, '-').replace('--', '-');
+  return createElement('h' + props.level, { id: slug }, props.children);
+};
+
 const Markdown: FunctionComponent<{
   body: string;
   version?: string;
 }> = function (props) {
   return (
-    <ReactMarkdownWithHtml
-      source={props.body}
-      escapeHtml={false}
+    <ReactMarkdown
+      children={props.body}
+      rehypePlugins={[rehypeRaw]}
       plugins={[gfm]}
-      renderers={{
-        code: CodeHighlighter,
+      components={{
+        code: ({ node, inline, className, children, ...props }) => {
+          const match = /language-(\w+)/.exec(className || '');
+          return !inline && match ? (
+            <SyntaxHighlighter
+              style={vscDarkPlus}
+              language={match[1]}
+              PreTag='div'
+              children={String(children).replace(/\n$/, '')}
+              {...props}
+            />
+          ) : (
+            <code className={className} {...props}>
+              {children}
+            </code>
+          );
+        },
+        img: ({ alt, src }) => {
+          return (
+            <img
+              alt={alt as string}
+              src={src as string}
+              className='img-fluid'
+            />
+          );
+        },
+        table: ({ children }) => <table className='table'>{children}</table>,
         blockquote: (content) => {
-          const value = content?.node?.children?.[0]?.children?.[0]?.value;
-
-          if (value && value.includes('##quote##')) {
-            return <Quote quote={JSON.parse(value)}></Quote>;
+          const value = content?.node?.children?.[1]?.children?.[0]?.value;
+          if (value && value.includes('##quotation##')) {
+            return <Quotation quotation={JSON.parse(value)}></Quotation>;
           } else {
-            return <Blockquote content={content.children}></Blockquote>;
+            return <Quote content={content.children}></Quote>;
           }
         },
-        heading: (props) => {
-          var children = Children.toArray(props.children);
-          var text = children.reduce(flatten, '');
-          var slug = text
-            .toLowerCase()
-            .trim()
-            .replace(/\W/g, '-')
-            .replace('--', '-');
-          return createElement('h' + props.level, { id: slug }, props.children);
-        }
+        h1: heading,
+        h2: heading,
+        h3: heading,
+        h4: heading,
+        h5: heading,
+        h6: heading
       }}
       transformLinkUri={transformLinkUri(props.version)}
       linkTarget={linkTarget}
