@@ -1,19 +1,7 @@
-import { FunctionComponent } from 'react';
-import { QuoteData } from '../models/common.model';
+import { FunctionComponent, ReactNode } from 'react';
 
-const Quote: FunctionComponent<{ quote: QuoteData }> = function ({ quote }) {
-  return (
-    <div className='quote'>
-      {quote.citation}
-      <div className='quote-metadata'>
-        <img className='quote-picture' src={`/images/${quote.pictureUrl}`} />
-        <div className='quote-author'>
-          <div>{quote.author}</div>
-          <div>{quote.authorRole}</div>
-        </div>
-      </div>
-    </div>
-  );
+const Quote: FunctionComponent<{ content: ReactNode }> = function (props) {
+  return <div className='alert quote'>{props.content}</div>;
 };
 
 export default Quote;
