@@ -1,3 +1,3 @@
 /// <reference types="next" />
 /// <reference types="next/types/global" />
-declare const ga: Function;
+/// <reference types="next/image-types/global" />
