@@ -1,6 +1,7 @@
 const version = require('./package.json').version;
+const withTM = require('next-transpile-modules')(['react-github-btn']);
 
-module.exports = {
+module.exports = withTM({
   generateBuildId: async () => {
     return `v${version}`;
   },
@@ -15,5 +16,6 @@ module.exports = {
     });
 
     return config;
-  }
-};
+  },
+  webpack5: true
+});
