@@ -1,3 +1,4 @@
+import Link from 'next/link';
 import { FunctionComponent } from 'react';
 import Hero from '../components/hero';
 import Meta from '../components/meta';
@@ -11,108 +12,108 @@ const SponsorUs: FunctionComponent = function () {
         description='Mockoon is an open source project built by a team of passionate volunteer maintainers.'
       />
 
-      <Hero title='Meet the team' />
+      <Hero
+        title='Meet the team'
+        subtitle='&#8220;We want to craft the best API tools on the market!&#8221;'
+      />
 
       <div className='container'>
-        <div className='columns'>
-          <div className='column is-6 is-offset-3'>
-            <div className='content has-text-centered'>
-              <p className='has-text-grey has-text-weight-semibold is-size-4 is-italic'>
-                &#8220;We want to craft the best API tools on the market!&#8221;
-              </p>
-            </div>
-          </div>
-        </div>
-        <div className='columns'>
-          <div className='column is-3 is-offset-3'>
-            <div className='section'>
-              <div className='content is-flex is-justify-content-center'>
-                <figure className='image is-128x128'>
+        <div className='row d-flex justify-content-center'>
+          <div className='col-6 col-lg-4'>
+            <div className='card card-row p-5'>
+              <div className='row gx-0 d-flex align-items-center justify-content-center flex-column'>
+                <div className='col-12 col-lg-7'>
                   <img
-                    className='is-rounded'
                     src='/images/about/guillaume.jpg'
                     alt='Mockoon maintainer'
+                    className='img-fluid rounded-circle mb-3 img-thumbnail shadow'
                   />
-                  <span className='title is-size-5'>Guillaume</span>
-                  <br />
-                  <a
-                    href='https://github.com/255kb'
-                    target='_blank'
-                    rel='noopener'
-                    className='has-text-grey'
-                  >
-                    <i className='icon-github'></i> 255kb
-                  </a>
-                </figure>
+                </div>
+                <div className='col-12 order-lg-1 d-flex justify-content-center'>
+                  <div className='card-meta pb-0 d-flex flex-column flex-lg-row'>
+                    <h6 className='text-uppercase text-muted me-lg-2 mb-lg-0 mb-1'>
+                      Guillaume
+                    </h6>
+
+                    <a
+                      href='https://github.com/255kb'
+                      target='_blank'
+                      rel='noopener'
+                      className='text-decoration-none'
+                    >
+                      <i className='icon-github fs-3 text-muted'></i>
+                    </a>
+                  </div>
+                </div>
               </div>
             </div>
           </div>
-          <div className='column is-3'>
-            <div className='section'>
-              <div className='content is-flex is-justify-content-center'>
-                <figure className='image is-128x128'>
+          <div className='col-6 col-lg-4'>
+            <div className='card card-row p-5'>
+              <div className='row gx-0 d-flex align-items-center justify-content-center flex-column'>
+                <div className='col-12 col-lg-7'>
                   <img
-                    className='is-rounded'
                     src='/images/about/fabrice.jpg'
                     alt='Mockoon maintainer'
+                    className='img-fluid rounded-circle mb-3 img-thumbnail shadow'
                   />
-                  <span className='title is-size-5'>Fabrice</span>
-                  <br />
-                  <a
-                    href='https://github.com/fabhoarau'
-                    target='_blank'
-                    rel='noopener'
-                    className='has-text-grey'
-                  >
-                    <i className='icon-github'></i> fabhoarau
-                  </a>
-                </figure>
+                </div>
+                <div className='col-12 order-lg-1 d-flex justify-content-center'>
+                  <div className='card-meta pb-0 d-flex flex-column flex-lg-row'>
+                    <h6 className='text-uppercase text-muted me-lg-2 mb-lg-0 mb-1'>
+                      Fabrice
+                    </h6>
+
+                    <a
+                      href='https://github.com/fabhoarau'
+                      target='_blank'
+                      rel='noopener'
+                      className='text-decoration-none'
+                    >
+                      <i className='icon-github fs-3 text-muted'></i>
+                    </a>
+                  </div>
+                </div>
               </div>
             </div>
           </div>
         </div>
-        <div className='columns'>
-          <div className='column is-6 is-offset-3'>
-            <div className='section'>
-              <div className='content'>
-                <div className='has-text-centered'></div>
-                <p>
-                  Mockoon is a free and open-source mock API tool created by
-                  Guillaume in 2017. Passionate about APIs and great developer
-                  tools, he spent a lot of time making Mockoon the easiest to
-                  use and quickest mock API tool.
-                </p>
-                <p>
-                  Recently joined by Fabrice, a long-time friend and full-stack
-                  developer, together they plan to bring Mockoon to the next
-                  level and deliver the best developer experience on the market.
-                </p>
-                <p>
-                  They believe in simplicity and efficiency and work hard every
-                  day to bring you a tool that is easy to set up yet powerful.
-                  The goal is to save you tons of time, not waste it.
-                </p>
-                <p>
-                  If you believe in what we are building, you can sponsor us in
-                  this journey and join the dozens of{' '}
-                  <a
-                    href='https://github.com/mockoon/mockoon/blob/main/backers.md'
-                    target='_blank'
-                    rel='noopener'
-                  >
-                    backers and supporters
-                  </a>
-                  .
-                </p>
-              </div>
-
-              <div className='has-text-centered'>
-                <a href='/sponsor-us/' className='button is-link is-light'>
-                  View sponsorship options
-                </a>
-              </div>
-            </div>
+        <div className='row d-flex justify-content-center align-items-center flex-column pt-4 pb-8 py-lg-8'>
+          <div className='col-lg-10 col-12 text-center text-lg-start'>
+            <p>
+              Mockoon is a free and open-source mock API tool created by
+              Guillaume in 2017. Passionate about APIs and great developer
+              tools, he spent a lot of time making Mockoon the easiest to use
+              and quickest mock API tool.
+            </p>
+            <p>
+              Recently joined by Fabrice, a long-time friend and full-stack
+              developer, together they plan to bring Mockoon to the next level
+              and deliver the best developer experience on the market.
+            </p>
+            <p>
+              They believe in simplicity and efficiency and work hard every day
+              to bring you a tool that is easy to set up yet powerful. The goal
+              is to save you tons of time, not waste it.
+            </p>
+            <p>
+              If you believe in what we are building, you can sponsor us in this
+              journey and join the dozens of{' '}
+              <a
+                href='https://github.com/mockoon/mockoon/blob/main/backers.md'
+                target='_blank'
+                rel='noopener'
+              >
+                backers and supporters
+              </a>
+              .
+            </p>
           </div>
+          <Link href='/sponsor-us/'>
+            <button className='col-10 col-lg-4 mt-4 btn btn-primary'>
+              View sponsorship options
+            </button>
+          </Link>
         </div>
       </div>
     </Layout>
