@@ -1,6 +1,4 @@
 import Cards from '../../components/cards';
-import ContactBanner from '../../components/contact-banner';
-import Download from '../../components/download';
 import Hero from '../../components/hero';
 import Meta from '../../components/meta';
 import Layout from '../../layout/layout';
@@ -29,14 +27,17 @@ export default function CaseStudies(props: { articles: ArticleList }) {
     <Layout>
       <Meta title={meta.title} description={meta.description} />
       <Hero title={meta.title} subtitle={meta.description} />
-      <Download />
 
-      <div className='section'>
+      <section className='pb-8'>
         <div className='container'>
-          <Cards small={true} path='case-studies' articles={props.articles} />
+          <Cards
+            large
+            path='case-studies'
+            articles={props.articles}
+            cover={false}
+          />
         </div>
-      </div>
-      <ContactBanner />
+      </section>
     </Layout>
   );
 }
