@@ -1,5 +1,4 @@
 import { FunctionComponent } from 'react';
-import ContactBanner from '../components/contact-banner';
 import Hero from '../components/hero';
 import Meta from '../components/meta';
 import SimpleCards from '../components/simple-cards';
@@ -31,71 +30,59 @@ const features: ItemCard = [
   }
 ];
 
-const SponsorUs: FunctionComponent = function () {
+const Cli: FunctionComponent = function () {
   return (
     <Layout>
       <Meta
         title='Take Mockoon to the command line'
         description='Mockoon CLI takes Mockoon where it has never been before: servers, CI environments, etc. Lightweight and fast.'
       />
-
       <Hero
-        title='Take Mockoon to the command line'
+        title="Take Mockoon to the <span class='text-primary'>command line</span>"
         subtitle='Mockoon CLI takes Mockoon where it has never been before. Lightweight and fast.'
-        cta={{
-          text: 'Installation instructions →',
-          link: 'https://github.com/mockoon/cli#installation'
-        }}
+        cta={[
+          {
+            text: 'Installation instructions →',
+            link: 'https://github.com/mockoon/cli#installation'
+          }
+        ]}
         mainPicture='/images/cli-screenshot.png'
+        mainPictureAlt='Mockoon CLI screenshot'
       />
-
-      <div className='container'>
-        <div className='section'>
-          <div className='columns'>
-            <div className='column'>
-              <div className='content'>
-                <h3 className='mb-6 is-size-4'>
-                  Mockoon's perfect complement for all your headless and
-                  automated environments.
-                </h3>
-              </div>
-              <SimpleCards items={features} />
+      <section className='py-5 py-lg-10'>
+        <div className='container'>
+          <div className='row py-5'>
+            <div className='text-center'>
+              <h2>
+                Mockoon's perfect complement for all your headless and automated
+                environments.
+              </h2>
+            </div>
+            <SimpleCards items={features} />
+          </div>
+          <div className='row py-5'>
+            <div className='col-md-6'>
+              <h3 className='text-center'>NPM package quick start</h3>
+              <Terminal
+                lines={[
+                  'npm install -g @mockoon/cli',
+                  'mockoon-cli start --data ./export-file.json'
+                ]}
+              />
+            </div>
+            <div className='col-md-6'>
+              <h3 className='text-center'>Docker image Quick start</h3>
+              <Terminal
+                lines={[
+                  'docker run -d --mount type=bind,source=/export-file.json,target=/data,readonly -p 3000:3000 mockoon/cli:latest -d data -i 0 -p 3000'
+                ]}
+              />
             </div>
           </div>
         </div>
-        <div className='section'>
-          <div className='columns'>
-            <div className='column'>
-              <div className='content'>
-                <h3 className='mb-6 is-size-4 has-text-centered'>
-                  NPM package quick start
-                </h3>
-                <Terminal
-                  lines={[
-                    'npm install -g @mockoon/cli',
-                    'mockoon-cli start --data ./export-file.json'
-                  ]}
-                />
-              </div>
-            </div>
-            <div className='column'>
-              <div className='content'>
-                <h3 className='mb-6 is-size-4 has-text-centered'>
-                  Docker image Quick start
-                </h3>
-                <Terminal
-                  lines={[
-                    'docker run -d --mount type=bind,source=/export-file.json,target=/data,readonly -p 3000:3000 mockoon/cli:latest -d data -i 0 -p 3000'
-                  ]}
-                />
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <ContactBanner />
+      </section>
     </Layout>
   );
 };
 
-export default SponsorUs;
+export default Cli;
