@@ -1,9 +1,7 @@
 import { FunctionComponent } from 'react';
 import Cards from '../../components/cards';
-import Download from '../../components/download';
 import Hero from '../../components/hero';
 import Meta from '../../components/meta';
-import Newsletter from '../../components/newsletter';
 import Layout from '../../layout/layout';
 import { ArticleList } from '../../models/common.model';
 import { buildIndexStaticProps } from '../../utils/static-builders';
@@ -37,14 +35,12 @@ const MockSamples: FunctionComponent<{
       `}</style>
       <Meta title={meta.title} description={meta.description} />
       <Hero title={meta.title} subtitle={meta.description} />
-      <Download />
 
-      <div className='section'>
+      <section className='pb-8'>
         <div className='container'>
-          <Cards path='mock-samples' articles={props.articles} />
+          <Cards path='mock-samples' articles={props.articles} medium />
         </div>
-      </div>
-      <Newsletter />
+      </section>
     </Layout>
   );
 };
