@@ -2,11 +2,7 @@ import matter from 'gray-matter';
 import Head from 'next/head';
 import React from 'react';
 import Article from '../../components/article';
-import ContactBanner from '../../components/contact-banner';
-import Download from '../../components/download';
-import Hero from '../../components/hero';
 import Meta from '../../components/meta';
-import Share from '../../components/share';
 import Layout from '../../layout/layout';
 import { ArticleData } from '../../models/common.model';
 import { buildSlugStaticPaths } from '../../utils/static-builders';
@@ -46,52 +42,19 @@ export default function Tutorial(props: {
         url={`/${props.slug}`}
         image={props.articleData.image}
       />
-
-      <Hero />
-
-      <div className='section'>
-        <div className='container'>
-          <div className='columns'>
-            <div className='column is-3'>
-              <div className='content'>
-                <Download />
-              </div>
-            </div>
-            <div className='column is-9'>
-              <Article
-                slug={props.slug}
-                path='tutorials'
-                articleBody={props.articleBody}
-                articleData={props.articleData}
-              />
-            </div>
-          </div>
+      <Article
+        slug={props.slug}
+        path='tutorials'
+        articleBody={props.articleBody}
+        articleData={props.articleData}
+      />
+      <section>
+        <div className='container text-center pt-3 pb-8'>
+          <a className='btn btn-secondary-soft' href='/tutorials/'>
+            ⬅ Back to the list of tutorials
+          </a>
         </div>
-        <div className='container'>
-          <div className='columns'>
-            <div className='column is-9 is-offset-3'>
-              <div className='content'>
-                <Share
-                  title='Find this tutorial useful? Share it!'
-                  url={`https://mockoon.com/${props.slug}/`}
-                  text={props.articleData.meta.description}
-                />
-              </div>
-            </div>
-          </div>
-        </div>
-        <div className='container'>
-          <div className='columns'>
-            <div className='column'>
-              <div className='content'>
-                <a href='/tutorials/'>⬅ Back to the tutorials</a>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <ContactBanner />
+      </section>
     </Layout>
   );
 }
